Add files via upload
An exploratory look at Covid 19 Numbers indicates that Covid was really kind to Africa, at least compared to the rest of the world.

Only 0.04% of her population were infected, the fewest across continents except Oceania – (a continent I have heard nothing about before this study) and with just over than 130,000 deaths, a little less than the total death recorded in the UK alone. Great statistics for Africa, considering that many of her Nations have below par health care facilities and poor crisis management abilities.

On the top of the infection ratio rung is South America with 6.8% of her population testing positive, a little more than their neighbors in the North with 6.6% and 0.2% more than Europe, which at the early stage of the virus, happen to be the epic Centre. Little surprise over 1 million people died of Covid in Europe, the highest across continents.

Asia recorded the most number of infected people, but this cannot be seen as a true reflection of the pandemic in Asia. Majority of these numbers are contributed by India – the only Asian Country in top 10 Death count, and China coming a distant second. Interestingly, despite the high infection cases, they have one of the best recovery rates, recording significantly fewer deaths than Europe, South and North America.